changed default screen to player creation.  start game now goes to game board
